Resync v0.0.10 of service-catalog with origin

+++ cmd/service-catalog/go/src/github.com/kubernetes-incubator/service-catalog/pkg/rest/core/fake/rest_client.go
@@ -19,6 +19,7 @@ package fake
 import (
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"net/http"
@@ -26,6 +27,7 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
+	scmeta "github.com/kubernetes-incubator/service-catalog/pkg/api/meta"
 	sc "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog"
 	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/testapi"
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -104,13 +106,13 @@ type RESTClient struct {
 }
 
 // NewRESTClient returns a new FakeCoreRESTClient
-func NewRESTClient() *RESTClient {
+func NewRESTClient(newEmptyObj func() runtime.Object) *RESTClient {
 	storage := make(NamespacedStorage)
 	watcher := NewWatcher()
 
 	coreCl := &fakerestclient.RESTClient{
 		Client: fakerestclient.CreateHTTPClient(func(request *http.Request) (*http.Response, error) {
-			r := getRouter(storage, watcher)
+			r := getRouter(storage, watcher, newEmptyObj)
 			rw := newResponseWriter()
 			r.ServeHTTP(rw, request)
 			return rw.getResponse(), nil
@@ -169,7 +171,11 @@ func (rw *responseWriter) getResponse() *http.Response {
 	}
 }
 
-func getRouter(storage NamespacedStorage, watcher *Watcher) http.Handler {
+func getRouter(
+	storage NamespacedStorage,
+	watcher *Watcher,
+	newEmptyObj func() runtime.Object,
+) http.Handler {
 	r := mux.NewRouter()
 	r.StrictSlash(true)
 	r.HandleFunc(
@@ -178,7 +184,7 @@ func getRouter(storage NamespacedStorage, watcher *Watcher) http.Handler {
 	).Methods("GET")
 	r.HandleFunc(
 		"/apis/servicecatalog.k8s.io/v1alpha1/namespaces/{namespace}/{type}",
-		createItem(storage),
+		createItem(storage, newEmptyObj),
 	).Methods("POST")
 	r.HandleFunc(
 		"/apis/servicecatalog.k8s.io/v1alpha1/namespaces/{namespace}/{type}/{name}",
@@ -186,7 +192,7 @@ func getRouter(storage NamespacedStorage, watcher *Watcher) http.Handler {
 	).Methods("GET")
 	r.HandleFunc(
 		"/apis/servicecatalog.k8s.io/v1alpha1/namespaces/{namespace}/{type}/{name}",
-		updateItem(storage),
+		updateItem(storage, newEmptyObj),
 	).Methods("PUT")
 	r.HandleFunc(
 		"/apis/servicecatalog.k8s.io/v1alpha1/namespaces/{namespace}/{type}/{name}",
@@ -234,11 +240,15 @@ func getItems(storage NamespacedStorage) func(http.ResponseWriter, *http.Request
 			// in memory, so we're going to make a deep copy first.
 			objCopy, err := conversion.NewCloner().DeepCopy(obj)
 			if err != nil {
-				log.Fatalf("error performing deep copy: %s", err)
+				errStr := fmt.Sprintf("error performing deep copy: %s", err)
+				http.Error(rw, errStr, http.StatusInternalServerError)
+				return
 			}
 			item, ok := objCopy.(runtime.Object)
 			if !ok {
-				log.Fatalf("error performing type assertion: %s", err)
+				errStr := fmt.Sprintf("error performing type assertion: %s", err)
+				http.Error(rw, errStr, http.StatusInternalServerError)
+				return
 			}
 			items = append(items, item)
 		}
@@ -249,61 +259,82 @@ func getItems(storage NamespacedStorage) func(http.ResponseWriter, *http.Request
 		case "brokers":
 			list = &sc.BrokerList{TypeMeta: newTypeMeta("broker-list")}
 			if err := meta.SetList(list, items); err != nil {
-				log.Fatalf("Error setting list items (%s)", err)
+				errStr := fmt.Sprintf("Error setting list items (%s)", err)
+				http.Error(rw, errStr, http.StatusInternalServerError)
+				return
 			}
 			codec, err = testapi.GetCodecForObject(&sc.BrokerList{})
 		case "serviceclasses":
 			list = &sc.ServiceClassList{TypeMeta: newTypeMeta("service-class-list")}
 			if err := meta.SetList(list, items); err != nil {
-				log.Fatalf("Error setting list items (%s)", err)
+				errStr := fmt.Sprintf("Error setting list items (%s)", err)
+				http.Error(rw, errStr, http.StatusInternalServerError)
+				return
 			}
 			codec, err = testapi.GetCodecForObject(&sc.ServiceClassList{})
 		case "instances":
 			list = &sc.InstanceList{TypeMeta: newTypeMeta("instance-list")}
 			if err := meta.SetList(list, items); err != nil {
-				log.Fatalf("Error setting list items (%s)", err)
+				errStr := fmt.Sprintf("Error setting list items (%s)", err)
+				http.Error(rw, errStr, http.StatusInternalServerError)
+				return
 			}
 			codec, err = testapi.GetCodecForObject(&sc.InstanceList{})
 		case "bindings":
 			list = &sc.BindingList{TypeMeta: newTypeMeta("binding-list")}
 			if err := meta.SetList(list, items); err != nil {
-				log.Fatalf("Error setting list items (%s)", err)
+				errStr := fmt.Sprintf("Error setting list items (%s)", err)
+				http.Error(rw, errStr, http.StatusInternalServerError)
+				return
 			}
 			codec, err = testapi.GetCodecForObject(&sc.BindingList{})
 		default:
-			log.Fatalf("unrecognized resource type: %s", tipe)
+			errStr := fmt.Sprintf("unrecognized resource type: %s", tipe)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		if err != nil {
-			log.Fatalf("error getting codec: %s", err)
+			errStr := fmt.Sprintf("error getting codec: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		listBytes, err := runtime.Encode(codec, list)
 		if err != nil {
-			log.Fatalf("error encoding list: %s", err)
+			errStr := fmt.Sprintf("error encoding list: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		rw.Write(listBytes)
 	}
 }
 
-func createItem(storage NamespacedStorage) func(rw http.ResponseWriter, r *http.Request) {
+func createItem(storage NamespacedStorage, newEmptyObj func() runtime.Object) func(rw http.ResponseWriter, r *http.Request) {
 	return func(rw http.ResponseWriter, r *http.Request) {
 		ns := mux.Vars(r)["namespace"]
 		tipe := mux.Vars(r)["type"]
-		// TODO: Is there some type-agnostic way to get the codec?
-		codec, err := testapi.GetCodecForObject(&sc.Broker{})
+		codec, err := testapi.GetCodecForObject(newEmptyObj())
 		if err != nil {
-			log.Fatalf("error getting codec: %s", err)
+			errStr := fmt.Sprintf("error getting a codec for %#v (%s)", newEmptyObj(), err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		bodyBytes, err := ioutil.ReadAll(r.Body)
 		if err != nil {
-			log.Fatalf("error getting body bytes: %s", err)
+			errStr := fmt.Sprintf("error getting body bytes: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		item, err := runtime.Decode(codec, bodyBytes)
 		if err != nil {
-			log.Fatalf("error decoding body bytes: %s", err)
+			errStr := fmt.Sprintf("error decoding body bytes: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		name, err := accessor.Name(item)
 		if err != nil {
-			log.Fatalf("couldn't get object name: %s", err)
+			errStr := fmt.Sprintf("couldn't get object name: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		if storage.Get(ns, tipe, name) != nil {
 			rw.WriteHeader(http.StatusConflict)
@@ -311,10 +342,13 @@ func createItem(storage NamespacedStorage) func(rw http.ResponseWriter, r *http.
 		}
 		accessor.SetResourceVersion(item, "1")
 		storage.Set(ns, tipe, name, item)
+		setContentType(rw)
 		rw.WriteHeader(http.StatusCreated)
 		bytes, err := runtime.Encode(codec, item)
 		if err != nil {
-			log.Fatalf("error encoding item: %s", err)
+			errStr := fmt.Sprintf("error encoding item (%s)", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		rw.Write(bytes)
 	}
@@ -332,17 +366,22 @@ func getItem(storage NamespacedStorage) func(http.ResponseWriter, *http.Request)
 		}
 		codec, err := testapi.GetCodecForObject(item)
 		if err != nil {
-			log.Fatalf("error getting codec: %s", err)
+			errStr := fmt.Sprintf("error getting codec (%s)", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		bytes, err := runtime.Encode(codec, item)
 		if err != nil {
-			log.Fatalf("error encoding item: %s", err)
+			errStr := fmt.Sprintf("error encoding item (%s)", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
+		setContentType(rw)
 		rw.Write(bytes)
 	}
 }
 
-func updateItem(storage NamespacedStorage) func(http.ResponseWriter, *http.Request) {
+func updateItem(storage NamespacedStorage, newEmptyObj func() runtime.Object) func(http.ResponseWriter, *http.Request) {
 	return func(rw http.ResponseWriter, r *http.Request) {
 		ns := mux.Vars(r)["namespace"]
 		tipe := mux.Vars(r)["type"]
@@ -352,26 +391,35 @@ func updateItem(storage NamespacedStorage) func(http.ResponseWriter, *http.Reque
 			rw.WriteHeader(http.StatusNotFound)
 			return
 		}
-		// TODO: Is there some type-agnostic way to get the codec?
-		codec, err := testapi.GetCodecForObject(&sc.Broker{})
+		codec, err := testapi.GetCodecForObject(newEmptyObj())
 		if err != nil {
-			log.Fatalf("error getting codec: %s", err)
+			errStr := fmt.Sprintf("error getting codec: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		bodyBytes, err := ioutil.ReadAll(r.Body)
 		if err != nil {
-			log.Fatalf("error getting body bytes: %s", err)
+			errStr := fmt.Sprintf("error getting body bytes: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		item, err := runtime.Decode(codec, bodyBytes)
 		if err != nil {
-			log.Fatalf("error decoding body bytes: %s", err)
+			errStr := fmt.Sprintf("error decoding body bytes: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
-		origResourceVersionStr, err := accessor.ResourceVersion(origItem)
+		origResourceVersionStr, err := accessor.ResourceVersion(item)
 		if err != nil {
-			log.Fatalf("error getting resource version")
+			errStr := fmt.Sprintf("error getting resource version")
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		resourceVersionStr, err := accessor.ResourceVersion(item)
 		if err != nil {
-			log.Fatalf("error getting resource version")
+			errStr := fmt.Sprintf("error getting resource version")
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
 		// As with the actual core apiserver, "0" is a special resource version that
 		// forces an update as if the current / most up-to-date resource version had
@@ -383,11 +431,56 @@ func updateItem(storage NamespacedStorage) func(http.ResponseWriter, *http.Reque
 		resourceVersion, err := strconv.Atoi(origResourceVersionStr)
 		resourceVersion++
 		accessor.SetResourceVersion(item, strconv.Itoa(resourceVersion))
-		storage.Set(ns, tipe, name, item)
+
+		// if the deletion timestamp is set and there are 0 finalizers, then we should delete
+		finalizers, err := scmeta.GetFinalizers(item)
+		if err != nil {
+			errStr := fmt.Sprintf("error getting finalizers (%s)", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
+		}
+		oldDT, err := scmeta.GetDeletionTimestamp(origItem)
+		if err != nil && err != scmeta.ErrNoDeletionTimestamp {
+			errStr := fmt.Sprintf("error getting deletion timestamp on existing obj (%s)", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
+		}
+		newDT, err := scmeta.GetDeletionTimestamp(item)
+		if err != nil && err != scmeta.ErrNoDeletionTimestamp {
+			errStr := fmt.Sprintf("error getting deletion timestamp on new obj (%s)", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
+		}
+		if newDT != nil && oldDT != nil {
+			// only measure the deletion timestamp based on seconds (.Unix()) rather than nanos
+			// (.UnixNano()), because nanoseconds are stored in memory but do not come over
+			// the wire. Thus, the new deletion timestamp will not be equal to the old
+			// because the new will have nanos missing
+			if newDT.Unix() != oldDT.Unix() {
+				errStr := fmt.Sprintf(
+					"you cannot update the deletion timestamp (old: %#v, new: %#v)",
+					oldDT.String(),
+					newDT.String(),
+				)
+				http.Error(rw, errStr, http.StatusBadRequest)
+				return
+			}
+		}
+
+		if len(finalizers) == 0 && newDT != nil {
+			// if there are no finalizers and the deletion timestamp is set, delete
+			storage.Delete(ns, tipe, name)
+		} else {
+			// otherwise, just update as normal
+			storage.Set(ns, tipe, name, item)
+		}
 		bytes, err := runtime.Encode(codec, item)
 		if err != nil {
-			log.Fatalf("error encoding item: %s", err)
+			errStr := fmt.Sprintf("error encoding item: %s", err)
+			http.Error(rw, errStr, http.StatusInternalServerError)
+			return
 		}
+		setContentType(rw)
 		rw.Write(bytes)
 	}
 }
@@ -402,8 +495,31 @@ func deleteItem(storage NamespacedStorage) func(http.ResponseWriter, *http.Reque
 			rw.WriteHeader(http.StatusNotFound)
 			return
 		}
-		storage.Delete(ns, tipe, name)
-		rw.WriteHeader(http.StatusAccepted)
+		finalizers, err := scmeta.GetFinalizers(item)
+		if err != nil {
+			http.Error(
+				rw,
+				fmt.Sprintf("error getting finalizers (%s)", err),
+				http.StatusInternalServerError,
+			)
+			return
+		}
+		if len(finalizers) == 0 {
+			// delete if there are no finalizers
+			storage.Delete(ns, tipe, name)
+		} else {
+			// set a deletion timestamp on the item if there are finalizers
+			if err := scmeta.SetDeletionTimestamp(item, time.Now()); err != nil {
+				http.Error(
+					rw,
+					fmt.Sprintf("error setting deletion timestamp (%s)", err),
+					http.StatusInternalServerError,
+				)
+				return
+			}
+			storage.Set(ns, tipe, name, item)
+		}
+		rw.WriteHeader(http.StatusOK)
 	}
 }
 
@@ -415,6 +531,7 @@ func listNamespaces(storage NamespacedStorage) func(http.ResponseWriter, *http.R
 				ObjectMeta: metav1.ObjectMeta{Name: ns},
 			})
 		}
+		setContentType(rw)
 		if err := json.NewEncoder(rw).Encode(&nsList); err != nil {
 			log.Printf("Error encoding namespace list (%s)", err)
 			rw.WriteHeader(http.StatusInternalServerError)
